added locking for libaio copy
